Ranamed db package to database. Moved tests from database package to models package for better coverage calculation.
