+++ database/database.go
@@ -1,5 +1,5 @@
 //nolint:lll
-package db
+package database
 
 import (
 	"context"
